+++ Library/Aphid/UI/ListViewItem.js
@@ -7,7 +7,7 @@
 
 Aphid.UI.ListViewItem = Class.create(Aphid.UI.View, {
 
-  displayName: "ListViewItem",
+  displayName: "Aphid.UI.ListViewItem",
 
   /**
    * Aphid.UI.ListViewItem#isSelected -> Boolean
@@ -46,7 +46,7 @@ Aphid.UI.ListViewItem = Class.create(Aphid.UI.View, {
     $super(options);
     if (!this.element)
     {
-      $L.debug("Initializing default element...", this.displayName)
+      $L.debug("Initializing default element...", this);
       this.element = new Element('li').addClassName("ListViewItem");
       this.isLoaded = true;
     }
@@ -68,7 +68,7 @@ Aphid.UI.ListViewItem = Class.create(Aphid.UI.View, {
   **/
   select: function()
   {
-    $L.debug("Selected...", this.displayName);
+    $L.debug("Selected...", this);
     this.element.addClassName('selected');
     this.isSelected = true;
     return this;
@@ -83,7 +83,7 @@ Aphid.UI.ListViewItem = Class.create(Aphid.UI.View, {
   **/
   deselect: function()
   {
-    $L.debug("Deselected...", this.displayName);
+    $L.debug("Deselected...", this);
     this.element.removeClassName('selected');
     this.isSelected = false;
     return this;